Version 0.8.1
Added new events for the UD Abadon armor sets (Warrior, Scout, and Witch)
Added new MCM option to filter out UD Abadon armor sets from normal device distribution
The Magic Key and Tattoo Removal Charm now work on non-player actors
The Frigid Wash now lowers players libedo by a configurable ammount when used (Requires OSL Aroused)
Added new Follower event settings

Added new Madness Plug events
Changed how the Madness Plug runs events requiring a notable effect from each event
Madness Plug events will no longer affect followers if allowFollowerEvents is false
the Madness Plug ending will now be random and not a full tie

Sex triggers now filter out actors other than the player based on the restraints they are wearing
Spells are now correctly unequipped when the player is put in heavy restraints
Fixed how Magic Key spawning works so they will spawn at the correct frequency
Magic Key now has an option (default false) to remove Oppressive Devices
Fixed bug where message boxes were not being displayed correctly or were showing ERROR
Made some slight optimization changes to improve performance

+++ src/tats.hpp
@@ -426,7 +426,7 @@ namespace DCURSES {
 									scriptManager.StartSex(target, false);
 								}
 								else {
-									scriptManager.ModifyArousal(target, static_cast<int>(mag / 10));
+									scriptManager.ModifyArousal(target, mag / 10);
 								}
 								break;
 							}
@@ -532,7 +532,7 @@ namespace DCURSES {
 				log::trace("events marks update");
 				switch (mark) {
 				case MARK::TAT_HEAT: {
-					scriptManager.ModifyArousal(player, settings.LMHeatMod / 4);
+					scriptManager.ModifyArousal(player, settings.LMHeatMod / 4.0f);
 					if (GetEffectMagnitude(HEAT_EFFECT) <= 0) {
 						RemoveLewdMark();
 						AIContextRemoveLewdMark();
@@ -550,7 +550,7 @@ namespace DCURSES {
 							for (auto const& actorHandle : *arr) {
 								auto actorPtr = actorHandle.get();
 								if (auto actor = actorPtr.get(); actor && actor->Is3DLoaded() && !actor->IsDead() && actor->GetPosition().GetDistance(playerPosition) <= settings.sexSearchRadius) {
-									scriptManager.ModifyArousal(actor, settings.LMAllureMod / 4);
+									scriptManager.ModifyArousal(actor, settings.LMAllureMod / 4.0f);
 								}
 							}
 						}
